Remove special WASI handling
I originally added special handling for WASI OS strings because I saw
`std::os::wasi::ffi::OsStrExt` in the docs and assumed that meant WASI
worked like Unix.

It doesn't. The consensus is to only permit Unicode. `wasmtime`
rejects invalid UTF-8:

  $ wasmtime target/wasm32-wasip2/debug/examples/echo.wasm $'\xff'
  Error: failed to convert "\xFF" to utf-8

Most platforms re-export Unix's `OsStrExt` and WASI happens to be one
of the platforms for which docs are built. But in practice any string
that passes between a WASI program and the OS must be Unicode. Invalid
OS strings can only be created using the extension traits.

Also, on the new `wasm32-wasip2` target `std::os::wasi::ffi::OsStrExt`
is marked unstable, and so lexopt failed to build.

So we no longer treat WASI specially in the library. It's now just
another fallback target that sometimes has to use
`OsStr::to_string_lossy`.

In the tests however we can now use `wasm32-wasip1` to finally test
how the fallback code reacts to invalid OS strings. This was never
very exciting but it's nice to have it covered.